functions/vision-ocr-test/requirements.txt
functions-framework==3.5.0
google-cloud-vision==3.4.4
requests==2.31.0 